refactor: move components to their own files and folders
