+++ packages/app-explorer/src/systems/Transaction/component/TxScripts/ReceiptItem/constants.ts
@@ -1,0 +1,6 @@
+import { createContext } from 'react';
+import type { ReceiptItemProps } from './types';
+
+export const TxScriptsContext = createContext<ReceiptItemProps>(
+  {} as ReceiptItemProps,
+);
