+++ packages/app-explorer/src/systems/Transaction/component/TxScripts/ReceiptItem/types.ts
@@ -1,0 +1,8 @@
+import type { GQLOperationReceipt, Maybe } from '@fuel-explorer/graphql/sdk';
+import type { BaseProps } from '@fuels/ui';
+
+export type ReceiptItemProps = BaseProps<{
+  receipt?: Maybe<GQLOperationReceipt>;
+  isIndented?: boolean;
+  hasPanic?: boolean;
+}>;
