+++ packages/app-explorer/src/systems/Transaction/component/TxScripts/ReceiptItemR/types.ts
@@ -1,0 +1,6 @@
+import type { GQLOperationReceipt } from '@fuel-explorer/graphql';
+
+export interface ReceiptItemRProps {
+  receipts?: GQLOperationReceipt[];
+  hasPanic: boolean;
+}
