+++ packages/app-explorer/src/systems/Transaction/component/TxScripts/TxReceiptBadge/constants.ts
@@ -1,0 +1,3 @@
+import { GQLReceiptType } from '@fuel-explorer/graphql/sdk';
+
+export const RETURN_TYPES = [GQLReceiptType.Return, GQLReceiptType.ReturnData];
