+++ packages/app-explorer/src/systems/Transaction/component/TxScripts/TxReceiptBlock/TxReceiptBlock.tsx
@@ -1,0 +1,24 @@
+import { useContext } from 'react';
+import { useMeasure } from 'react-use';
+
+import { Collapsible, ScrollArea } from '@fuels/ui';
+import { JsonViewer } from '~/systems/Core/components/JsonViewer/JsonViewer';
+import { TxScriptsContext } from '~/systems/Transaction/component/TxScripts/ReceiptItem/constants';
+import { parseTXScriptJson } from '~/systems/Transaction/component/TxScripts/utils';
+import { styles } from './styles';
+
+export function TxReceiptBlock() {
+  const { receipt } = useContext(TxScriptsContext);
+  const classes = styles();
+  const [ref, { width }] = useMeasure();
+  return (
+    <Collapsible.Content
+      ref={ref as React.Ref<HTMLDivElement>}
+      className={classes.utxos()}
+    >
+      <ScrollArea style={{ width }}>
+        <JsonViewer data={parseTXScriptJson(receipt?.item)} />
+      </ScrollArea>
+    </Collapsible.Content>
+  );
+}
