+++ packages/app-explorer/src/systems/Transaction/component/TxScripts/TypesCounter/types.ts
@@ -1,0 +1,8 @@
+import {
+  GQLTransactionReceiptFragment,
+  Maybe,
+} from '@fuel-explorer/graphql/sdk';
+
+export interface TypesCounterProps {
+  receipts?: Maybe<GQLTransactionReceiptFragment[]>;
+}
